refactor table to only include generic funcs and move specific show functions into respecitve command modules
